bug, feat: Fixed packing fraction bug and improved CLI
The previous commit had issues to pack to very high pressures, due to
the parameters and adaptation mechanisms being too lax. Now we are
passing the parameters as the user needs them through the CLI, such as
the minimum translation, deformation and volume changes. We are also
exposing the frequency to adjust these parameters.